refactor(app-wrapper): replace scss with styled component
