+++ src/components/app-wrapper/app-wrapper.spec.js
@@ -2,7 +2,7 @@ import React from "react";
 import TestUtils from "react-dom/test-utils";
 import { shallow } from "enzyme";
 import { rootTagTest } from "../../utils/helpers/tags/tags-specs";
-import AppWrapper from "./app-wrapper";
+import AppWrapper from "./app-wrapper.component";
 
 describe("app wrapper", () => {
   let instance;
@@ -14,26 +14,28 @@ describe("app wrapper", () => {
   });
 
   it("renders the children", () => {
-    let div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
+    const div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
     expect(div.textContent).toEqual("foo");
   });
 
   it("renders with correct classes", () => {
-    let div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
-    expect(div.className).toEqual("carbon-app-wrapper foobar");
+    const div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
+    expect(div.className).toContain("carbon-app-wrapper foobar");
   });
 
   it("renders with additional html attributes", () => {
     instance = TestUtils.renderIntoDocument(
       <AppWrapper style={{ color: "red" }}>foo</AppWrapper>
     );
-    let div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
+    const div = TestUtils.findRenderedDOMComponentWithTag(instance, "div");
     expect(div.style.color).toEqual("red");
   });
 
   describe("tags", () => {
     describe("on component", () => {
-      let wrapper = shallow(<AppWrapper data-element="bar" data-role="baz" />);
+      const wrapper = shallow(
+        <AppWrapper data-element="bar" data-role="baz" />
+      );
 
       it("include correct component, element and role data tags", () => {
         rootTagTest(wrapper, "app-wrapper", "bar", "baz");
