+++ src/components/app-wrapper/app-wrapper.style.js
@@ -1,9 +1,9 @@
-@import "./../../style-config/general";
+import styled from "styled-components";
 
-.carbon-app-wrapper {
+const StyledAppWrapper = styled.div`
   box-sizing: border-box;
-  max-width: $app-max-width;
-  min-width: $app-width;
+  max-width: 1600px;
+  min-width: 958px;
   margin: 0 auto;
   padding: 0 40px;
 
@@ -14,4 +14,6 @@
   @media only screen and (max-width: 1024px) {
     padding: 0 25px;
   }
-}
+`;
+
+export default StyledAppWrapper;
