+++ src/components/app-wrapper/index.js
@@ -1,0 +1,1 @@
+export { default } from "./app-wrapper.component";
