--- src/components/app-wrapper/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "./app-wrapper.js"
-}
